% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomposeVar.R
\name{.detectFeatureIdType}
\alias{.detectFeatureIdType}
\title{Detect type of feature identifier}
\usage{
.detectFeatureIdType(featureType, genes, M, maxMissing = 10000, verbose = TRUE)
}
\arguments{
\item{featureType}{Character scalar, one of \code{"AUTO"},
\code{"ENSEMBL_GENE_ID"}, \code{"GENE_SYMBOL"}, \code{"ENTREZ_GENE_ID"}
or \code{"ARCHS4_ID"}.}

\item{genes}{Gene annotation data frame.}

\item{M}{Expression matrix (feature identifiers are in row names).}

\item{maxMissing}{Numeric scalar. This function will throw an error if more
than \code{maxMissing} features in \code{genes} are not available in
\code{M}.}

\item{verbose}{Logical scalar. If \code{TRUE}, report on progress.}
}
\value{
\code{featureType} as a character scalar.
}
\description{
Detect type of feature identifier
}
\author{
Panagiotis Papasaikas, Michael Stadler
}
\keyword{internal}
