% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.grid_cor_wNAs}
\alias{.grid_cor_wNAs}
\title{Calculate correlation between a numeric matrix and a (large)
HDF5MAtrix/DelayedMatrix using grid access. Both Matrices can contain NAs.
If \code{thr} is specified, substitution of values less than \code{thr} with
\code{NA}s will be performed on the HDF5Matrix}
\usage{
.grid_cor_wNAs(query, hdf5, chunk_size = 1000, workers = 16, thr = NULL)
}
\arguments{
\item{query}{Numeric matrix n x k.}

\item{hdf5}{HDF5Matrix/DelayedMatrix n x l, where l is typically >> k}

\item{chunk_size}{column dimension for the grid used to read blocks from the
HDF5 Matrix. Should be larger than/equal to the ncol chunkdim used to write
the data on disk.}

\item{workers}{Number of workers used for parallelization}

\item{thr}{If specified, a low bound on expression. Values lower than that
are substituted by \code{NA}s on the HDF5 Matrix}
}
\value{
Correlation matrix k x l
}
\description{
Calculate correlation between a numeric matrix and a (large)
HDF5MAtrix/DelayedMatrix using grid access. Both Matrices can contain NAs.
If \code{thr} is specified, substitution of values less than \code{thr} with
\code{NA}s will be performed on the HDF5Matrix
}
\author{
Panagiotis Papasaikas
}
\keyword{internal}
