% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.grid_cor_woNAs}
\alias{.grid_cor_woNAs}
\title{Calculate correlation between a numeric matrix and a (large)
HDF5MAtrix/DelayedMatrix using grid access. It is assumed that both Matrices
do not contain \code{NA}s.}
\usage{
.grid_cor_woNAs(
  query,
  hdf5,
  chunk_size = 1000,
  BPPARAM = BiocParallel::bpparam()
)
}
\arguments{
\item{query}{Numeric matrix n x k.}

\item{hdf5}{HDF5Matrix/DelayedMatrix n x l, where l is typically >> k}

\item{chunk_size}{column dimension for the grid used to read blocks from the
HDF5 Matrix. Should be larger than/equal to the ncol chunkdim used to write
the data on disk.}

\item{BPPARAM}{BiocParallelParam object specifying how parallelization is to be performed.}
}
\value{
Correlation matrix k x l
}
\description{
Calculate correlation between a numeric matrix and a (large)
HDF5MAtrix/DelayedMatrix using grid access. It is assumed that both Matrices
do not contain \code{NA}s.
}
\author{
Panagiotis Papasaikas
}
\keyword{internal}
