% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryWithContrasts.R
\name{.loadContrastDatabase}
\alias{.loadContrastDatabase}
\title{Load contrast database}
\usage{
.loadContrastDatabase(organism = c("Human", "Mouse"))
}
\arguments{
\item{organism}{Character scalar selecting the organism for which to load the
contrast database. One of \code{"Human"} or \code{"Mouse"}.}
}
\value{
A \code{SummarizedExperiment} with pre-calculated contrasts as
    assays.
}
\description{
Load a pre-calculated, organism-specific contrast database and return it
as a \code{SummarizedExperiment}.
}
\author{
Panagiotis Papasaikas, Michael Stadler
}
\keyword{internal}
