% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.M2Mcor}
\alias{.M2Mcor}
\title{Calculate correlation between corresponding columns of two matrices}
\usage{
.M2Mcor(M1, M2)
}
\arguments{
\item{M1, M2}{Numeric matrices of dimensions n x k.}
}
\value{
Vector of length k with correlations between corresponding columns.
}
\description{
Calculate correlation between corresponding columns of two matrices
}
\author{
Panagiotis Papasaikas, Charlotte Soneson
}
\keyword{internal}
