% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQueryResults.R
\name{.plotManhDens}
\alias{.plotManhDens}
\title{Visualize query results as a composite Manhattan/Density plot}
\usage{
.plotManhDens(scores, annot = "")
}
\arguments{
\item{scores}{Numeric named vector (typically of z-scores) to use for
visualization}

\item{annot}{Annotation dataframe for the topn results to highlight}
}
\value{
A composite manhattan/density plot
}
\description{
Visualize query results as a composite Manhattan/Density plot
}
\examples{
 qRES <- queryWithContrasts(MyDecomposedContrasts)
.plotManhDens(qRES$zscores[[1]][1,], qRES$TopHits[[contrast]][[dset]] ) 



}
\author{
Panagiotis Papasaikas
}
