% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomposeVar.R
\name{.preprocessInput}
\alias{.preprocessInput}
\title{Preprocess count matrix}
\usage{
.preprocessInput(M, ids, verbose, pseudocount = 4)
}
\arguments{
\item{M}{Numeric matrices with feature counts (features are in rows,
samples in columns).}

\item{ids}{Character vector. Only rows of \code{M} with row names in
\code{ids} will be returned.}

\item{verbose}{Logical scalar. Report on progress if \code{TRUE}.}

\item{pseudocount}{Numeric scalar, added to the scaled (per million) counts
before log2-transformation.}
}
\value{
Scaled, log2-transformed and subset version of \code{M}.
}
\description{
The given count matrix \code{M} will be preprocessed by:
library normalization (calculation of counts per million, using the
column sums as library sizes), log2-transformed after addition of
\code{pseudocount}, and \code{NA} values will be set to 0.
The subset of \code{M} with row names in \code{ids} is returned.
}
\author{
Panagiotis Papasaikas, Michael Stadler
}
\keyword{internal}
