% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthos.R
\docType{package}
\name{orthos-package}
\alias{orthos}
\alias{orthos-package}
\title{orthos - variance decomposition using conditional variational auto-encoders}
\description{
\pkg{orthos} decomposes RNA-seq contrasts, for example obtained from a
gene knock-out or compound treatment experiment, into unspecific and
experiment-specific components. Original and decomposed contrasts can
be efficiently queried against a large database of publiclycontrasts
(derived from ARCHS4, https://maayanlab.cloud/archs4/) to identify similar
experiments. \pkg{orthos} furthermore provides plotting functions to
visualize the results of such a search for similar contrasts.
}
\author{
\strong{Maintainer}: Panagiotis Papasaikas \email{panagiotis.papasaikas@fmi.ch} (\href{https://orcid.org/0000-0002-1640-7636}{ORCID})

Authors:
\itemize{
  \item Charlotte Soneson \email{charlotte.soneson@fmi.ch} (\href{https://orcid.org/0000-0003-3833-2169}{ORCID})
  \item Michael Stadler \email{michael.stadler@fmi.ch} (\href{https://orcid.org/0000-0002-2269-4934}{ORCID})
}

Other contributors:
\itemize{
  \item Friedrich Miescher Institute for Biomedical Research [copyright holder]
}

}
