% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQueryResults.R
\name{plotQueryResults}
\alias{plotQueryResults}
\title{Plot query results}
\usage{
plotQueryResults(scores, topn = 10)
}
\arguments{
\item{topn}{}
}
\description{
Plot query results
}
\author{
Panagiotis Papasaikas
}
