% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryWithContrasts_DA.R
\name{queryWithContrasts_DA}
\alias{queryWithContrasts_DA}
\title{Query the contrast database with a set of contrasts}
\usage{
queryWithContrasts_DA(
  contrasts = NULL,
  use = c("expressed.in.both", "all.genes"),
  exprThr = 0.25,
  organism = c("Human", "Mouse"),
  plotContrast = c("RESIDUAL", "INPUT", "DECODED", "NONE"),
  detailTopn = 10,
  verbose = TRUE,
  workers = round(4 + parallel::detectCores()/4),
  chunk_size = 500
)
}
\arguments{
\item{contrasts}{A \code{SummarizedExperiment} object with assays containing 
contrasts named INPUT_CONTRASTS, DECODED_CONTRASTS and RESIDUAL_CONTRASTS 
(at least one should be present) and context information in an assay 
named CONTEXT. The latter is only required when use="expressed.in.both".
This is typically generated using  
\code{decomposeVar}.}

\item{use}{Determines if all.genes or genes expressed in both query and 
target context will be used. Note that "expressed.in.both", though more 
accurate, is slower.}

\item{exprThr}{is the quantile in the provided context that determines the 
expression value above which a gene is considered to be expressed. This 
same value is then used for thresholding the contrast database. Only 
applies when use="expressed.in.both"}

\item{detailTopn}{specifies the number of top hits for which metadata will 
be returned in the TopHits slot of the results.}

\item{verbose}{Logical scalar indicating whether to print messages along 
the way.}

\item{workers}{Number of workers used for parallelizetion (passed to BiocParallel::MulticoreParam).}

\item{chunk_size}{Column dimension for the grid used to read blocks from the HDF5 Matrix.
Sizes between 250 and 1000 are recommended. Smaller sizes reduce memory usage.}
}
\value{
A list of PearsonRhos, Zscores against the datbase as well as 
    detailed Metadata for the detailTopn hits.
}
\description{
Query the contrast database with a set of contrasts
}
\author{
Panagiotis Papasaikas
}
