% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdeJUNKERenv.R
\name{testdeJUNKERenv}
\alias{testdeJUNKERenv}
\title{Test conda environment}
\usage{
testdeJUNKERenv()
}
\value{
A list indicating whether keras is available, and the version
    of TensorFlow.
}
\description{
Test conda environment
}
\examples{
testdeJUNKERenv()

}
\author{
Charlotte Soneson
}
