% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testorthosenv.R
\name{testOrthosEnv}
\alias{testOrthosEnv}
\title{Test conda environment}
\usage{
testOrthosEnv()
}
\value{
A list indicating whether keras is available, and the version
    of TensorFlow.
}
\description{
Test conda environment
}
\examples{
\donttest{
testOrthosEnv()
}

}
\author{
Charlotte Soneson
}
