% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomposeVar.R
\name{decomposeVar}
\alias{decomposeVar}
\title{Decompose input contrasts to decoded and residual fractions}
\usage{
decomposeVar(
  M,
  MD = NULL,
  treatm = NULL,
  cntr = NULL,
  processInput = TRUE,
  organism = c("Human", "Mouse"),
  featureType = c("AUTO", "ENSEMBL_GENE_ID", "GENE_SYMBOL", "ENTREZ_GENE_ID",
    "ARCHS4_ID"),
  verbose = TRUE
)
}
\arguments{
\item{M}{Matrix of raw gene counts.}

\item{MD}{Matrix of gene deltas (optional). If \code{MD} is specified,
\code{M} is assumed to be a raw gene count matrix specifying context for
contrasts specified in \code{MD}. \code{MD} is then a matrix of gene
deltas with the same dimensions as \code{M}. If \code{MD} is specified,
\code{treatm} and \code{cntr} have to be \code{NULL}.}

\item{treatm, cntr}{Vectors indicating column indices in \code{M}
corresponding to treatments and controls. If \code{treatm} and
\code{cntr} are specified, \code{MD} has to be \code{NULL}.}

\item{processInput}{If set to \code{TRUE} (default) the count matrix will 
be preprocessed (library normalized, log-transformed, NA values will 
be set to 0).}

\item{organism}{Selects the autoencoder model trained on data from this
species. One of \code{"Human"} or \code{"Mouse"}.}

\item{featureType}{Set to \code{"AUTO"} for automatic feature id-type
detection. Alternatively specify the type of supplied id features.
Current supported types are \code{"ENSEMBL_GENE_ID"},
\code{"GENE_SYMBOL"}, \code{"ENTREZ_GENE_ID"} and \code{"ARCHS4_ID"}.}

\item{verbose}{Logical scalar indicating whether to print messages along 
the way.}
}
\value{
A \code{\link[SummarizedExperiment]{SummarizedExperiment}} object
    with the decomposed contrasts in the assays and the decomposed variance
    as the \code{\link[SummarizedExperiment]{colData}}.
}
\description{
Decompose input contrasts (gene expression Deltas) to decoded (generic) and residual (unique) components
according to a contrast encoder-decoder pre-trained on a large corpus of public RNAseq experiments.
}
\examples{
\dontrun{
RES <- decomposeVar(M = ContextM, MD = DeltaM, processInput = FALSE)
RES <- decomposeVar(M = CountM, treatm = c(3, 4, 5, 6), 
                    cntr = c(1, 1, 2, 2), processInput = FALSE)
}

}
\author{
Panagiotis Papasaikas
}
