% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryWithContrasts.R
\name{.loadContrastDatabase}
\alias{.loadContrastDatabase}
\title{Load contrast database}
\usage{
.loadContrastDatabase(
  organism = c("Human", "Mouse"),
  mode = c("ANALYSIS", "DEMO")
)
}
\arguments{
\item{organism}{Character scalar selecting the organism for which to load the
contrast database. One of \code{"Human"} or \code{"Mouse"}.}

\item{mode}{When in "ANALYSIS" mode (default) the complete contrast DB is queried. "DEMO" mode
employs a small "toy" database for the queries. "DEMO" should only be used for testing/demonstration purposes
and never for actual analysis purposes.}
}
\value{
A \code{SummarizedExperiment} with pre-calculated contrasts as
    assays.
}
\description{
Load a pre-calculated, organism-specific contrast database and return it
as a \code{SummarizedExperiment}.
}
\author{
Panagiotis Papasaikas, Michael Stadler
}
\keyword{internal}
