% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQueryResults.R
\name{plotQueryResults}
\alias{plotQueryResults}
\title{Plot query results}
\usage{
plotQueryResults(scores, topn = 10, annot = "")
}
\arguments{
\item{scores}{Numeric named vector (typically of z-scores) to use for
visualization}

\item{topn}{Highlight the topn results}

\item{annot}{Annotation vector used for coloring the topn results}
}
\value{
A plot
}
\description{
Plot query results
}
\examples{
x <- rnorm(100)
plotQueryResults(scores = x,
                 topn = 10,
                 annot = as.character(round(abs(x), 1)))

}
\author{
Panagiotis Papasaikas
}
