% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQueryResults.R
\name{plotQueryResultsViolin}
\alias{plotQueryResultsViolin}
\title{Visualize query results as violin plots}
\usage{
plotQueryResultsViolin(query.results, plot = TRUE)
}
\arguments{
\item{query.results}{A list containing the results of a query performed with
\code{queryWithContrasts}}

\item{plot}{Logical scalar specifying if a plot should be generated.}
}
\value{
A list of ggplot violin plots (one for each dataset) for the scores
of queries using different contrast components against the respective
contrast DBs.
}
\description{
Visualize query results as violin plots
}
\examples{
\dontrun{
qRES <- queryWithContrasts(MyDecomposedContrasts)
plotQueryResults.violin(qRES)
}

}
\author{
Panagiotis Papasaikas
}
