% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryWithContrasts.R
\name{queryWithContrasts}
\alias{queryWithContrasts}
\title{Query the contrast database with a set of contrasts}
\usage{
queryWithContrasts(
  contrasts = NULL,
  use = c("expressed.in.both", "all.genes"),
  exprThr = 0.25,
  organism = "Human",
  preserveInGlobalEnv = TRUE,
  plotContrast = c("RESIDUAL", "INPUT", "DECODED", "NONE"),
  detailTopn = 10
)
}
\arguments{
\item{contrasts}{A \code{SummarizedExperiment} object with assays containing 
contrasts named INPUT_CONTRASTS, DECODED_CONTRASTS and RESIDUAL_CONTRASTS 
(at least one should be present) and context information in an assay 
named CONTEXT. The latter is only required when use="expressed.in.both".
This is typically generated using  
\code{decomposeVar}.}

\item{use}{Determines if all.genes or genes expressed in both query and 
target context will be used. Note that "expressed.in.both", though more 
accurate, is much slower.}

\item{exprThr}{is the quantile in the provided context that determines the 
expression value above which a gene is considered to be expressed. This 
same value is then used for thresholding the contrast database. Only 
applies when use="expressed.in.both"}

\item{preserveInGlobalEnv}{specifies whether the contrast database (stored 
as a \code{SummarizedExperiment} object in \code{target.contrasts})
should be preserved in the GlobalEnv either for future queries or for 
accessing its metadata}

\item{detailTopn}{specifies the number of top hits for which metadata will 
be returned in the TopHits slot of the results.}
}
\value{
A list of PearsonRhos, Zscores against the datbase as well as 
    detailed Metadata for the detailTopn hits.
}
\description{
Query the contrast database with a set of contrasts
}
\author{
Panagiotis Papasaikas
}
