% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdeJUNKERenv.R
\name{testdeJUNKERenv}
\alias{testdeJUNKERenv}
\title{Test conda environment}
\usage{
testdeJUNKERenv()
}
\value{
Nothing is returned
}
\description{
Test conda environment
}
\examples{
testdeJUNKERenv()

}
\author{
Charlotte Soneson
}
